feat(sys): windows backend

+++ src/sys/mod.rs
@@ -12,13 +12,10 @@ use unix as os;
 #[cfg(windows)]
 use windows as os;
 
-// Global pointers to the AtomicBools provided by the user via Arc.
-// These are private to the `sys` module so only the handlers can see them.
 pub(crate) static SIGINT_PTR: AtomicPtr<AtomicBool> = AtomicPtr::new(std::ptr::null_mut());
 pub(crate) static SIGTERM_PTR: AtomicPtr<AtomicBool> = AtomicPtr::new(std::ptr::null_mut());
 pub(crate) static SIGQUIT_PTR: AtomicPtr<AtomicBool> = AtomicPtr::new(std::ptr::null_mut());
 
-/// The unified entry point for registering a flag.
 pub fn register_flag(sig: TermSignal, flag: *mut AtomicBool) -> HadesResult<()> {
     match sig {
         TermSignal::SIGINT => SIGINT_PTR.store(flag, Ordering::Release),
@@ -29,17 +26,14 @@ pub fn register_flag(sig: TermSignal, flag: *mut AtomicBool) -> HadesResult<()> 
     os::register_signal(sig)
 }
 
-/// Initializes the signals backend (e.g. self-pipe on Unix).
-pub fn setup_signals_backend() -> HadesResult<i32> {
+pub fn setup_signals_backend() -> HadesResult<isize> {
     os::setup_pipe()
 }
 
-/// Blocks until a signal is received via the backend.
-pub fn wait_for_signal(backend_handle: i32) -> HadesResult<u8> {
+pub fn wait_for_signal(backend_handle: isize) -> HadesResult<u8> {
     os::read_signal(backend_handle)
 }
 
-/// Registers multiple signals at once for the backend.
 pub fn register_signals(signals: &[TermSignal]) -> HadesResult<()> {
     for &sig in signals {
         os::register_signal(sig)?;
